// ss-front/src/app/model/ImageFields.ts
export interface ImageFields {
  upperJawOcclusalView: string | ArrayBuffer | null;
  intraoralFrontalView: string | ArrayBuffer | null;
  lowerJawOcclusalView: string | ArrayBuffer | null;
  rightSideLateralView: string | ArrayBuffer | null;
  leftSideLateralView: string | ArrayBuffer | null;
  leftSideView: string | ArrayBuffer | null;
  frontalView: string | ArrayBuffer | null;
  rightSideView: string | ArrayBuffer | null;
}